requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.26
pydantic>=2.5
pyyaml>=6.0
